Comp tests fixed
