+++ src/main/java/com/github/loki4j/common/LokiResponse.java
